version 1.4.1:
  - Fixed a bug in the initialization configuration
  - Modify the method widget of class Yod_Controller
  - Add a constant YOD_LIBPATH (default value is YOD_RUNPATH)
  - Add a constant YOD_DATADIR to replace constant YOD_LOGPATH
  - Modify the method model and autoload of class Yod_Base
  - Modify the method getInstance of class Yod_Model
  - Modify the method __construct and handle of class Yod_Server
  - Fixed a bug in yod_request_route (multi params analysis is abnormal)
  - Fixed a bug in yod_controller_render (scope of variable)
  - Add a protected property _fields in the class Yod_Model
  - Modify the method find and select of class Yod_Model
  - Modify the method find and select of class Yod_DbModel
  - Fixed a bug in yod_controller_widget
  - Fixed a bug in yod_controller_construct
  - Fixed a bug in yod_request_dispatch
  - Add a method session in the class Yod_Base
  - Add public method queryData and queryRows in the class Yod_Database
  - Modify the method __call of class Yod_Client
  - Fixed a bug in yod_database_* and yod_dbpdo_execute

+++ yod/yod_database.h
@@ -23,9 +23,9 @@ void yod_database_construct(yod_database_t *object, zval *config TSRMLS_DC);
 int yod_database_getinstance(zval *config, zval *result TSRMLS_DC);
 int yod_database_config(yod_database_t *object, char *name, uint name_len, zval *value, zval *retval TSRMLS_DC);
 int yod_database_dbconfig(yod_database_t *object, zval *config, long linknum, zval *retval TSRMLS_DC);
-int yod_database_insert(yod_database_t *object, zval *data, char *table, uint table_len, int replace, zval *retval TSRMLS_DC);
-int yod_database_update(yod_database_t *object, zval *data, char *table, uint table_len, char *where, uint where_len, zval *params, zval *retval TSRMLS_DC);
-int yod_database_delete(yod_database_t *object, char *table, uint table_len, char *where, uint where_len, zval *params, zval *retval TSRMLS_DC);
+ulong yod_database_insert(yod_database_t *object, zval *data, char *table, uint table_len, int replace, zval *retval TSRMLS_DC);
+ulong yod_database_update(yod_database_t *object, zval *data, char *table, uint table_len, char *where, uint where_len, zval *params, zval *retval TSRMLS_DC);
+ulong yod_database_delete(yod_database_t *object, char *table, uint table_len, char *where, uint where_len, zval *params, zval *retval TSRMLS_DC);
 int yod_database_select(yod_database_t *object, zval *select, char *table, uint table_len, char *where, uint where_len, zval *params, char *extend, uint extend_len, zval *retval TSRMLS_DC);
 
 extern zend_class_entry *yod_database_ce;
